+++ keyboards/idobao/id87/v1/readme.md
@@ -10,10 +10,10 @@ A TKL hotswap board from IDOBAO.
 
 Make example for this keyboard (after setting up your build environment):
 
-    make idobao/id87:default
+    make idobao/id87/v1:default
 
 Flashing example for this keyboard:
 
-    make idobao/id87:default:flash
+    make idobao/id87/v1:default:flash
 
 See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
