+++ keyboards/idobao/id87/v1/rules.mk
