+++ keyboards/idobao/id87/v1/v1.c
@@ -12,4 +12,4 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "id87.h"
+#include "v1.h"
